requirements.txt
# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
pydantic-settings==2.1.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database Clients
asyncpg==0.29.0
clickhouse-driver==0.2.9
clickhouse_connect==0.7.0
aiochclient==2.5.1
questdb==1.1.0
psycopg2-binary==2.9.9
psycopg==3.1.14
redis[hiredis]==5.0.1

# Message Queue
nats-py==2.6.0
asyncio-nats-streaming==0.4.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-prometheus==0.42b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-asyncpg==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0

# Utils
python-dotenv==1.0.0
tenacity==8.2.3
httpx==0.25.2
aiohttp==3.9.1
aiodns==3.1.1
charset-normalizer==3.4.1
orjson==3.9.10
ujson==5.8.0
email-validator==2.1.0.post1
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
pendulum==2.1.2
croniter==2.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0
pylint==3.0.2

# Message Queue Dependencies
celery[redis]==5.3.6
flower==2.0.1

# Database Dependencies
alembic==1.13.1

# Dagster Core Dependencies
dagster==1.9.11
dagster-graphql==1.9.11
dagster-webserver==1.9.11
dagster-postgres==0.25.11
dagster-celery[redis]==0.25.11
dagster-docker==0.25.11
dagster-pandas==0.25.11
dagster-prometheus==0.25.11
dagster-shell==0.25.11

# Additional dependencies for Dagster
python-dateutil==2.8.2
pytz==2023.3
psutil==5.9.8 